# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SkillRapido - Real-time service marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface MessageResponse {
  message: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  customer: "customer",
  provider: "provider",
} as const;

export interface RegisterBody {
  name: string;
  email: string;
  phone?: string;
  password: string;
  role: RegisterBodyRole;
  latitude?: number;
  longitude?: number;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserProfileRole =
  (typeof UserProfileRole)[keyof typeof UserProfileRole];

export const UserProfileRole = {
  customer: "customer",
  provider: "provider",
  admin: "admin",
} as const;

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  phone?: string | null;
  role: UserProfileRole;
  avatarUrl?: string | null;
  latitude?: number | null;
  longitude?: number | null;
  createdAt: string;
}

export interface AuthResponse {
  user: UserProfile;
  message?: string;
}

export interface Skill {
  id: number;
  name: string;
  category: string;
  icon: string;
  description?: string | null;
}

export interface ProviderProfile {
  id: number;
  userId: number;
  name: string;
  email: string;
  phone?: string | null;
  bio?: string | null;
  isOnline: boolean;
  isVerified: boolean;
  rating?: number | null;
  totalJobs: number;
  totalEarnings: number;
  skills: Skill[];
  latitude?: number | null;
  longitude?: number | null;
}

export interface UpdateProviderProfileBody {
  bio?: string;
  phone?: string;
  skillIds?: number[];
  latitude?: number;
  longitude?: number;
}

export interface AvailabilityBody {
  isOnline: boolean;
}

export interface LocationBody {
  latitude: number;
  longitude: number;
}

export interface EarningsSummary {
  totalEarnings: number;
  todayEarnings: number;
  weekEarnings: number;
  monthEarnings: number;
  pendingAmount: number;
  completedJobs: number;
  averageJobValue: number;
}

export interface CreateJobBody {
  title: string;
  description: string;
  skillId: number;
  budget: number;
  address: string;
  latitude: number;
  longitude: number;
  scheduledAt?: string | null;
}

export type JobStatus = (typeof JobStatus)[keyof typeof JobStatus];

export const JobStatus = {
  open: "open",
  assigned: "assigned",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Job {
  id: number;
  title: string;
  description: string;
  status: JobStatus;
  budget: number;
  commission: number;
  providerAmount: number;
  address: string;
  latitude: number;
  longitude: number;
  scheduledAt?: string | null;
  completedAt?: string | null;
  customerId: number;
  customerName: string;
  providerId?: number | null;
  providerName?: string | null;
  skill: Skill;
  bidCount: number;
  createdAt: string;
}

export type BidStatus = (typeof BidStatus)[keyof typeof BidStatus];

export const BidStatus = {
  pending: "pending",
  accepted: "accepted",
  rejected: "rejected",
} as const;

export interface Bid {
  id: number;
  jobId: number;
  providerId: number;
  providerName: string;
  providerRating?: number | null;
  providerTotalJobs: number;
  message?: string | null;
  status: BidStatus;
  createdAt: string;
}

export type JobWithBids = Job & {
  bids: Bid[];
};

export interface AcceptJobBody {
  message?: string;
}

export interface JobCompletionResult {
  job: Job;
  customerCharged: number;
  providerPaid: number;
  appCommission: number;
}

export interface CustomerDashboard {
  totalJobsPosted: number;
  openJobs: number;
  completedJobs: number;
  cancelledJobs: number;
  totalSpent: number;
  averageJobCost: number;
  recentJobs: Job[];
}

export interface ProviderDashboard {
  isOnline: boolean;
  totalEarnings: number;
  todayEarnings: number;
  completedJobs: number;
  activeJobs: number;
  rating?: number | null;
  pendingBids: number;
  recentJobs: Job[];
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  job_posted: "job_posted",
  job_accepted: "job_accepted",
  job_completed: "job_completed",
  provider_joined: "provider_joined",
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  message: string;
  timestamp: string;
  jobId?: number | null;
  skillName?: string | null;
}

export type GetProviderJobsParams = {
  /**
   * Filter by status
   */
  status?: GetProviderJobsStatus;
};

export type GetProviderJobsStatus =
  (typeof GetProviderJobsStatus)[keyof typeof GetProviderJobsStatus];

export const GetProviderJobsStatus = {
  active: "active",
  completed: "completed",
  all: "all",
} as const;

export type ListJobsParams = {
  status?: ListJobsStatus;
};

export type ListJobsStatus =
  (typeof ListJobsStatus)[keyof typeof ListJobsStatus];

export const ListJobsStatus = {
  open: "open",
  assigned: "assigned",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
  all: "all",
} as const;